Break up NodeSearch into smaller components
