+++ src/core/components/controls/NodeSearch/NodeListItem.tsx
@@ -1,0 +1,56 @@
+import React, { FC } from 'react';
+import NodeModel from '../../../NodeModel';
+
+interface Props {
+  node: NodeModel;
+  handleSelect: (event: any) => void;
+}
+
+export const NodeListItem: FC<Props> = ({
+  node,
+  handleSelect,
+}) => {
+  const elementDataProperties = {
+    id: node.name,
+    'data-node-model-variation-name': node.name,
+  };
+
+  // HOW TO ONLY ALLOW CLICK E FROM PARENT?
+  // REPEAT THE EDATA FOR ALL CHILDREN FOR NOW
+  return (
+    <li
+      key={node.category + node.name}
+      onDoubleClick={handleSelect}
+      {...elementDataProperties}
+      className="py-3 flex"
+      tabIndex={2}
+    >
+      <div className="ml-3">
+        <div
+          className="flex text-sm mb-2 font-medium text-gray-900 text-bold"
+          {...elementDataProperties}
+        >
+          <div
+            {...elementDataProperties}
+            className="text-indigo-500"
+          >
+            {node.category}
+          </div>
+          <div {...elementDataProperties} className="">
+            ::{node.name}
+          </div>
+        </div>
+        <div
+          className="text-xs text-gray-500"
+          {...elementDataProperties}
+        >
+          <span className="ml-2" {...elementDataProperties}>
+            {node.summary}
+          </span>
+        </div>
+      </div>
+    </li>
+  );
+};
+
+export default NodeListItem;
