+++ src/core/components/controls/NodeSearch/index.ts
@@ -1,0 +1,3 @@
+import NodeSearch from './NodeSearch';
+
+export default NodeSearch;
